application/requirements.txt
Pillow==8.1.2
flask==1.1.2
flask-cors==3.0.10
redis==3.5.3